updates to css, html, and javascript
Updated css file and html to be cleaner. Pulled two long sections of code out of the html and created separate js scripts called at the bottom. Also added a timer to the initial draw map to make sure the function does not try to run before the data are loaded.